Remove checks for active terms in services
Also stop using the Hyrax namespace.
