--- app/services/hyrax/partners_service.rb
@@ -1,8 +1,0 @@
-# services/partners_service.rb
-module Hyrax
-  class PartnersService < Hyrax::QaSelectService
-    def initialize
-      super('partnering_agencies')
-    end
-  end
-end
