+++ app/services/partners_service.rb
@@ -1,0 +1,6 @@
+# services/partners_service.rb
+class PartnersService < Hyrax::QaSelectService
+  def initialize
+    super('partnering_agencies')
+  end
+end
