+++ app/services/submitting_service.rb
@@ -1,0 +1,6 @@
+# services/submitting_service.rb
+class SubmittingService < Hyrax::QaSelectService
+  def initialize
+    super('submitting_type')
+  end
+end
