--- app/views/records/edit_fields/hyrax/_submitting_type.html.erb
@@ -1,6 +1,0 @@
-<% submitting_service = Hyrax::SubmittingService.new %>
-<%= f.input :submitting_type, as: :select,
-    collection: submitting_service.select_all_options,
-    include_blank: true,
-    item_helper: submitting_service.method(:include_current_value),
-    input_html: { class: 'form-control' } %>
